refactor(ActionableList): extract types to ActionableList.types.ts
