+++ src/components/molecules/ActionableList/index.tsx
@@ -1,2 +1,2 @@
-export type { IActionableListItem, IActionableListProps, IActionableListTexts } from "./ActionableList";
+export type { IActionableListItem, IActionableListProps, IActionableListTexts } from "./ActionableList.types";
 export { default } from "./ActionableList";
